openvm_stark_sdk::bench
Function run_with_metric_collection

pub fn run_with_metric_collection<R>(
    output_path_envar: impl AsRef<OsStr>,
    f: impl FnOnce() -> R,
) -> R

Run a function with metric collection enabled. The metrics will be written to a file specified by an environment variable which name is output_path_envar.