openvm_stark_sdk::config
Module baby_bear_poseidon2_root

Stark Config for root stark, which field is BabyBear but polynomials are committed in Bn254.

Structs

BabyBearPermutationRootEngine

Functions

config_from_perm

default_config

pcs_log_degree is the upper bound on the log_2(PCS polynomial degree).

default_engine

pcs_log_degree is the upper bound on the log_2(PCS polynomial degree).

engine_from_perm

print_hash_counts

Logs hash count statistics to stdout and returns as struct. Count of 1 corresponds to a Poseidon2 permutation with rate RATE that outputs OUT field elements

root_perm

The permutation for outer recursion.

Type Aliases

BabyBearPermutationRootConfig

BabyBearPoseidon2RootConfig

BabyBearPoseidon2RootEngine