openvm_stark_sdk::dummy_airs::interaction
Module dummy_interaction_air

Air with columns | count | fields[..] |

Chip will either send or receive the fields with multiplicity count. The main Air has no constraints, the only constraints are specified by the Chip trait

Structs

DummyInteractionAir

DummyInteractionChip

Note: in principle, committing cached trace is out of scope of a chip. But this chip is for usually testing, so we support it for convenience.

DummyInteractionCols

DummyInteractionData